デッキ関係のAPI

+++ src/main/resources/db/migration/V2__Create_tokens_table.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS tokens (
-    id              INT GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
-    user_id         INT NOT NULL REFERENCES users(id) ON DELETE CASCADE,
+    id              BIGINT GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
+    user_id         BIGINT NOT NULL REFERENCES users(id) ON DELETE CASCADE,
     access_token    UUID NOT NULL,
     refresh_token   UUID NOT NULL,
     created_at      TIMESTAMPTZ NOT NULL DEFAULT now()
